% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportReImToNMRium.R
\name{exportReIm}
\alias{exportReIm}
\title{Method for exporting spectra information into the NMRium format.}
\usage{
exportReIm(
  x,
  yr,
  yi,
  output = NA,
  observeFrequency = 600,
  dataType = "NMR SPECTRUM",
  solvent = "UNKNOWN",
  nucleus = "1H",
  col = "#000000",
  meta = list(PHC0 = 0, PHC1 = 0),
  info = list()
)
}
\arguments{
\item{x}{Independent variable}

\item{yr}{Real dependent variable}

\item{yi}{(optional)Imaginary dependent variable}

\item{output}{Object containing other NMRium data}

\item{observeFrequency}{Observe frequency in MHz}

\item{dataType}{Data type format. Use 'NMR FID' if the spectrum is in the time domain (Default 'NMR SPECTRUM')}

\item{solvent}{String that specify the solvent. (Default 'UNKNOWN')}

\item{nucleus}{Observe nucleus.  (Default '1H')}

\item{col}{Color for this spectrum (Default black)}

\item{meta}{A named list with elements that you want to include in the spectra meta i.e. PHC0, PHC1. Used for spectra processing}

\item{info}{A named list with elements that you want to include in the spectra info. i.e. PHC0, PHC1, name, AGE, SEX, etc. Used for visualization}
}
\value{
list
}
\description{
You can use this to export simple x, re and x, re and im spectra
If you provide a list containing NMRium data, it will append the new spectra data into the proper place,
so that you can get an overlay of several spectra into the platform.
}
