% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportReImToNMRium.R
\name{fillGapsWith}
\alias{fillGapsWith}
\title{Given a f(x) -> y, with x being a progression of real numbers, with the same delta on its domain, except for some gaps in the middle
this function returns a new pair of x2 and y2 such that all pairs in x,y are contained and x2 is a monotonic progression of real numbers
i.e a series of equally spaced data on x, with all the gaps filled with the given value.}
\usage{
fillGapsWith(x, y, value)
}
\arguments{
\item{x}{Independent variable}

\item{y}{Dependent variable}

\item{value}{A value to fill the gaps on y}
}
\value{
list(newX, newY)
}
\description{
Given a f(x) -> y, with x being a progression of real numbers, with the same delta on its domain, except for some gaps in the middle
this function returns a new pair of x2 and y2 such that all pairs in x,y are contained and x2 is a monotonic progression of real numbers
i.e a series of equally spaced data on x, with all the gaps filled with the given value.
}
