% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportReImToNMRium.R
\name{isEquallySpacedX}
\alias{isEquallySpacedX}
\title{Checks that x is a progression of real numbers with constant delta}
\usage{
isEquallySpacedX(x)
}
\arguments{
\item{x}{Independent variable}
}
\value{
boolean
}
\description{
Checks that x is a progression of real numbers with constant delta
}
