% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.NMRium.R
\name{nmriumView}
\alias{nmriumView}
\title{Method for exporting spectra information into the NMRium format.}
\usage{
nmriumView(
  nmriumData,
  srv,
  name = "default",
  title = "exported from R",
  nmriumURL = "https://www.nmrium.org/nmrium#?toc",
  openBrowser = TRUE,
  digits = 12
)
}
\arguments{
\item{nmriumData}{nmriumData}

\item{srv}{A server}

\item{name}{A name to give to the generated file. Don't include extension}

\item{title}{String to identify the element in NMRium}

\item{nmriumURL}{The NMRium instance}

\item{openBrowser}{It must open the browser to display the TOC?}
}
\value{
void
}
\description{
You can use this to export simple x, re and x, re and im spectra
If you provide a list containing NMRium data, it will append the new spectra data into the proper place,
so that you can get an overlay of several spectra into the platform.
}
