% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.NMRium.R
\name{nmriumViewToc}
\alias{nmriumViewToc}
\title{Method for exporting table of content containing spectra information into the NMRium format.}
\usage{
nmriumViewToc(
  toc,
  srv,
  name = "toc_view",
  nmriumURL = "https://www.nmrium.org/nmrium#?toc",
  openBrowser = TRUE
)
}
\arguments{
\item{toc}{A object describing a TOC. It can contain NMRium object or references to spectra files}

\item{srv}{A server}

\item{name}{A name for this toc. This avoid overwriting other toc that could being served}

\item{nmriumURL}{The NMRium instance}

\item{openBrowser}{It must open the browser to display the TOC?}
}
\value{
void
}
\description{
You can use this to export simple x, re and x, re and im spectra
If you provide a list containing NMRium data, it will append the new spectra data into the proper place,
so that you can get an overlay of several spectra into the platform.
}
